+++ src/main/java/net/meander/subtlyd/util/data/tags/BlockTagsSD.java
@@ -83,7 +83,7 @@ public class BlockTagsSD extends FabricTagsProvider.BlockTagsProvider {
         valueLookupBuilder(BlockTags.NEEDS_STONE_TOOL)
                 .add(BlocksSD.IRON_GRATE);
         valueLookupBuilder(BlockTags.MINEABLE_WITH_HOE)
-                .add(BlocksSD.WARPED_WART_OVERHANG);
+                .add(BlocksSD.WARPED_OVERHANG);
         valueLookupBuilder(TRIGGERS_AMBIENT_WIND_BLOCK_SOUNDS)
                 .add(Blocks.SNOW, Blocks.SNOW_BLOCK, Blocks.POWDER_SNOW)
                 .add(Blocks.ICE, Blocks.BLUE_ICE, Blocks.PACKED_ICE, Blocks.FROSTED_ICE)