+++ src/main/java/net/meander/subtlyd/world/block/BlocksSD.java
@@ -49,7 +49,7 @@ public class BlocksSD {
             .sound(SoundType.WET_GRASS)
             .offsetType(BlockBehaviour.OffsetType.XZ)
             .pushReaction(PushReaction.DESTROY));
-    public static final Block WARPED_WART_OVERHANG = register("warped_wart_overhang", WarpedWartOverhangBlock::new , BlockBehaviour.Properties.ofFullCopy(Blocks.WARPED_WART_BLOCK)
+    public static final Block WARPED_OVERHANG = register("warped_overhang", WarpedOverhangBlock::new , BlockBehaviour.Properties.ofFullCopy(Blocks.WARPED_WART_BLOCK)
             .noCollision()
             .noOcclusion()
             .instabreak()